refactor: add Help in fullscreen dashboard settings
