+++ src/components/dashboard/subsettings/Help.tsx
@@ -1,0 +1,5 @@
+import { HelpListItems } from "~components/help/HelpListItems";
+
+export function HelpDashboardView() {
+  return <HelpListItems />;
+}
