+++ src/routes/dashboard/dashboard.constants.ts
@@ -26,6 +26,7 @@ import { TokensDashboardView } from "~components/dashboard/Tokens";
 import { ContactsDashboardView } from "~components/dashboard/Contacts";
 import { NotificationSettingsDashboardView } from "~components/dashboard/NotificationSettings";
 import { AboutDashboardView } from "~components/dashboard/About";
+import { HelpDashboardView } from "~components/dashboard/subsettings/Help";
 
 // Advance Settings:
 import { SignSettingsDashboardView } from "~components/dashboard/SignSettings";
@@ -81,6 +82,13 @@ export const basicSettings: (DashboardRouteConfig | Setting)[] = [
     component: NotificationSettingsDashboardView,
   },
   getSetting("gateways"),
+  {
+    name: "help",
+    displayName: "help",
+    description: "help_description",
+    icon: HelpCircle,
+    component: HelpDashboardView,
+  },
   {
     name: "about",
     displayName: "setting_about",
